game quiz1 ver 1

+++ src/main/java/dictionary/controllers/SearchController.java
@@ -37,7 +37,6 @@ import java.net.URL;
 import java.util.ResourceBundle;
 
 import javafx.scene.control.Button;
-
 public class SearchController implements Initializable {
     @FXML
     private TextField searchBox = new TextField();
@@ -110,12 +109,15 @@ public class SearchController implements Initializable {
 
     @FXML
     public void handleEdit() {
+        if (searchBox.getText().isEmpty() || searchBox.getText().isBlank()) return;
+        if (wordToFind.getWord().isBlank() || wordToFind.getWord().isEmpty()) return;
         if (relatedResults.getItems().size() <=0){
             return;
         }
         Dialog<String> dialog = new Dialog<>();
-        dialog.setHeaderText(null);
-
+        dialog.setTitle("Edit word");
+        DialogPane tmp = dialog.getDialogPane();
+        tmp.getStylesheets().add(getClass().getResource("/style/dialog.css").toExternalForm());
         Label wordLabel = new Label("Word: ");
         Label display = new Label(wordToFind.getWord());
 
@@ -141,6 +143,7 @@ public class SearchController implements Initializable {
 
         dialog.getDialogPane().setContent(gridPane);
 
+
         ButtonType okButton = new ButtonType("OK", ButtonBar.ButtonData.OK_DONE);
         ButtonType cancelButton = new ButtonType("Cancel", ButtonBar.ButtonData.CANCEL_CLOSE);
         dialog.getDialogPane().getButtonTypes().addAll(okButton, cancelButton);
@@ -163,10 +166,13 @@ public class SearchController implements Initializable {
 
                 WordsDao.modifyWord(wordToFind.getWord(), "meaning", meaning, "anhviet");
                 Alert successAlert = new Alert(Alert.AlertType.INFORMATION);
+                DialogPane tmp1 = successAlert.getDialogPane();
+                tmp1.getStylesheets().add(getClass().getResource("/style/dialog.css").toExternalForm());
+                successAlert.setTitle("Success");
                 successAlert.setContentText("Word information updated successfully!");
                 successAlert.showAndWait();
                 wordDefinition.setText("Type:\n" + wordToFind.getType()+ "\nMeaning:\n" + wordToFind.getMeaning());
-
+                wordDisplay.setText(wordToFind.getWord() + "\n" + wordToFind.getPronunciation());
             }
             else {
                 dialog.close();
@@ -177,7 +183,11 @@ public class SearchController implements Initializable {
 
     @FXML
     public void handleDelete() {
+        if (searchBox.getText().isEmpty() || searchBox.getText().isBlank()) return;
         Alert alert = new Alert(Alert.AlertType.INFORMATION);
+        alert.setHeaderText("Add a new word");
+        DialogPane tmp = alert.getDialogPane();
+        tmp.getStylesheets().add(getClass().getResource("/style/dialog.css").toExternalForm());
         if(WordsDao.deleteWord(wordToFind.getWord(),"anhviet")) {
             alert.setContentText("delete successfully");
             alert.showAndWait();
@@ -186,6 +196,10 @@ public class SearchController implements Initializable {
     @FXML
     public void handleAdd() {
         Dialog<String> dialog = new Dialog<>();
+        dialog.setHeaderText("Add a new word");
+        DialogPane tmp = dialog.getDialogPane();
+        tmp.getStylesheets().add(getClass().getResource("/style/dialog.css").toExternalForm());
+        //tmp.getStyleClass().add("dialog");
         dialog.setHeaderText(null);
 
         Label newWordLabel = new Label("New word: ");
@@ -242,11 +256,17 @@ public class SearchController implements Initializable {
                 if (id != -1) {
                     System.out.println("This word already in dictionary");
                     Alert successAlert = new Alert(Alert.AlertType.INFORMATION);
+                    DialogPane tmp1 = successAlert.getDialogPane();
+                    tmp1.getStylesheets().add(getClass().getResource("/style/dialog.css").toExternalForm());
+                    successAlert.setTitle("Success");
                     successAlert.setContentText("This word already in dictionary");
                     successAlert.showAndWait();
                 }
                 else if (WordsDao.addWord(ye,"anhviet")) {
                     Alert successAlert = new Alert(Alert.AlertType.INFORMATION);
+                    DialogPane tmp1 = successAlert.getDialogPane();
+                    tmp1.getStylesheets().add(getClass().getResource("/style/dialog.css").toExternalForm());
+                    successAlert.setTitle("Success");
                     successAlert.setContentText("Word information updated successfully!");
                     successAlert.showAndWait();
                 }
@@ -258,6 +278,8 @@ public class SearchController implements Initializable {
     }
     @FXML
     public void handleAddCollection() {
+        if (searchBox.getText().isEmpty() || searchBox.getText().isBlank()) return;
+        if (wordToFind.getWord().isBlank() || wordToFind.getWord().isEmpty()) return;
         Dialog<String> dialog = new Dialog<>();
         dialog.setHeaderText(null);
 
@@ -270,6 +292,10 @@ public class SearchController implements Initializable {
         GridPane gridPane = new GridPane();
         gridPane.add(INSTRUCTION, 1, 1);
         gridPane.add(allCollections, 1, 2);
+
+        dialog.setHeaderText("Add this word into a collection");
+        DialogPane tmp = dialog.getDialogPane();
+        tmp.getStylesheets().add(getClass().getResource("/style/dialog.css").toExternalForm());
 
         dialog.getDialogPane().setContent(gridPane);
